Updated the hosting model to direclty host the Blazor WASM application from the main Umbraco Backend instance. Moved the IContentDeliveryService to the shared core library and added two different implementations depending on the server (SSR) context or the client (WASM) context.
Adding support for persisting pre-rendered content to prevent a flash on the UI when the client app binds to the server pre-rendered data.

Removing the controller to redirect to the back end since the root of
the back end now servers a SSR enabled version of the Blazor WASM
Client.